pyproject.toml
[tool.poetry]
name = "share-fair"
version = "0.1.0"
description = ""
authors = ["Devyansh Vakharia <devyanshvakharia2510@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
telethon = "^1.36.0"
python-dotenv = "^1.0.1"
fastapi = "^0.114.1"
google-generativeai = "^0.8.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
